requirements.txt
streamlit
python-dotenv
langchain
openai
langchain-community
langchain-openai
pandas
python-pptx
python-docx
